Split some circular dependencies
